requirements.txt
requests==2.22.0
watchdog==2.3.1
python-dotenv==0.20.0